refactor: Organize routes in routes directory
