+++ routes/404.js
@@ -1,0 +1,9 @@
+const router = require("express").Router();
+
+router.get("*", (req, res) => {
+  res.status(404).render("error", {
+    Error: { code: 404, status: "Not Found", message: "Page Not Found!" }
+  });
+});
+
+module.exports = router;
