+++ views/create.ejs
@@ -11,7 +11,7 @@
       <!-- Contact Thumbnail -->
       <section class="thumbnail-details">
         <section class="buttons">
-          <a href="/" id="cancel-btn"><p>Cancel</p></a>
+          <a href="/contact" id="cancel-btn"><p>Cancel</p></a>
           <a
             id="save-btn"
             onclick="document.querySelector('#submit-btn').click()"
